refactor: extract types into individual files and enhance tsconfig
- Create src/types/ directory with 12 individual type files
- Each interface/type now has its own file for maintainability
- Update imports in app.ts and index.ts to use ./types/index.js
- Remove old types.ts file
- tsconfig.build.json now extends tsconfig.json
- All 52 tests pass

Closes #5
