+++ src/types/appOptions.ts
@@ -1,0 +1,7 @@
+/**
+ * Configuration options for creating a Bunary app.
+ */
+export interface AppOptions {
+	/** Base path prefix for all routes (default: "") */
+	basePath?: string;
+}
