+++ src/types/httpMethod.ts
@@ -1,0 +1,4 @@
+/**
+ * HTTP method types supported by the router.
+ */
+export type HttpMethod = "GET" | "POST" | "PUT" | "DELETE" | "PATCH" | "HEAD" | "OPTIONS";
