+++ src/types/pathParams.ts
@@ -1,0 +1,4 @@
+/**
+ * Path parameters extracted from route patterns (e.g., `/users/:id` → `{ id: "123" }`).
+ */
+export type PathParams = Record<string, string>;
