+++ src/types/queryParams.ts
@@ -1,0 +1,4 @@
+/**
+ * Query parameters parsed from the URL search string.
+ */
+export type QueryParams = Record<string, string | string[]>;
